+++ src/main/java/dev/julian/ezshow/platform/forge112/EzShowForgeMod.java
@@ -21,7 +21,7 @@ import java.util.UUID;
 public final class EzShowForgeMod {
     public static final String MOD_ID = "ezshow";
     public static final String NAME = "ezshow";
-    public static final String VERSION = "1.1.1";
+    public static final String VERSION = "1.1.2";
 
     @EventHandler
     public void init(FMLInitializationEvent event) {